+++ completion_aggregator/settings/common.py
@@ -4,6 +4,8 @@ Common settings for completion_aggregator.
 
 from __future__ import absolute_import, division, print_function, unicode_literals
 
+from event_routing_backends.utils.settings import event_tracking_backends_config
+
 
 def plugin_settings(settings):
     """
@@ -42,3 +44,22 @@ def plugin_settings(settings):
     # 1. All courses should be reaggregated for the changes to take effect.
     # 2. It's not possible to revert this change by reaggregation without manually removing existing Aggregators.
     settings.COMPLETION_AGGREGATOR_AGGREGATE_UNRELEASED_BLOCKS = False
+
+    # Whitelist the aggregator events for use with event routing backends xAPI backend.
+    # If these settings don't already exist, then ERB hasn't been loaded yet, so we need to set them to empty lists.
+    # But once ERB does load it will append its events to our list, preserving what we added here.
+    if not hasattr(settings, 'EVENT_TRACKING_BACKENDS_ALLOWED_XAPI_EVENTS'):
+        settings.EVENT_TRACKING_BACKENDS_ALLOWED_XAPI_EVENTS = []
+    if not hasattr(settings, 'EVENT_TRACKING_BACKENDS_ALLOWED_CALIPER_EVENTS'):
+        settings.EVENT_TRACKING_BACKENDS_ALLOWED_CALIPER_EVENTS = []
+    enabled_aggregator_events = [
+        f'openedx.completion_aggregator.progress.{block_type}'
+
+        for block_type in settings.COMPLETION_AGGREGATOR_TRACKING_EVENT_TYPES
+    ]
+    settings.EVENT_TRACKING_BACKENDS_ALLOWED_XAPI_EVENTS += enabled_aggregator_events
+    settings.EVENT_TRACKING_BACKENDS.update(event_tracking_backends_config(
+        settings,
+        settings.EVENT_TRACKING_BACKENDS_ALLOWED_XAPI_EVENTS,
+        settings.EVENT_TRACKING_BACKENDS_ALLOWED_CALIPER_EVENTS,
+    ))