+++ test_settings.py
@@ -49,12 +49,13 @@ INSTALLED_APPS = (
     'django.contrib.auth',
     'django.contrib.sessions',
     'django.contrib.admin',
-    'completion_aggregator',
     'completion',
     'oauth2_provider',
     'waffle',
     'test_utils.test_app',
     'eventtracking.django.apps.EventTrackingConfig',
+    'event_routing_backends',
+    'completion_aggregator',
 )
 
 LOCALE_PATHS = [root('completion_aggregator', 'conf', 'locale')]
@@ -85,6 +86,11 @@ USE_TZ = True
 
 # Enables event tracking in the tests, see https://github.com/openedx/event-tracking
 EVENT_TRACKING_ENABLED = True
+EVENT_TRACKING_BACKENDS = {}
+# Provided so the generated xAPI events use a known LMS URL when testing.
+LMS_ROOT_URL = "http://localhost:18000"
+# Provided so the generated xAPI events use a known "event routing backends package string" when testing.
+RUNNING_WITH_TEST_SETTINGS = True
 
 # pylint: disable=unused-import,wrong-import-position
 from test_utils.test_app import celery  # isort:skip